Move files to package structure
