+++ src/test/kotlin/no/ssb/metadata/vardef/controllers/VariableDefinitionByIdControllerTest.kt
@@ -1,4 +1,4 @@
-package no.ssb.metadata.vardef
+package no.ssb.metadata.vardef.controllers
 
 import io.micronaut.http.HttpStatus
 import io.restassured.http.ContentType
