+++ src/test/kotlin/no/ssb/metadata/vardef/controllers/VariableDefinitionsControllerTest.kt
@@ -1,4 +1,4 @@
-package no.ssb.metadata.vardef
+package no.ssb.metadata.vardef.controllers
 
 import io.micronaut.http.HttpStatus
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest
