+++ src/test/kotlin/no/ssb/metadata/vardef/models/VariableDefinitionTest.kt
@@ -1,9 +1,5 @@
-package no.ssb.metadata.vardef
+package no.ssb.metadata.vardef.models
 
-import no.ssb.metadata.vardef.models.CompleteResponse
-import no.ssb.metadata.vardef.models.RenderedVariableDefinition
-import no.ssb.metadata.vardef.models.SavedVariableDefinition
-import no.ssb.metadata.vardef.models.SupportedLanguages
 import no.ssb.metadata.vardef.utils.COMPLETE_RESPONSE
 import no.ssb.metadata.vardef.utils.INCOME_TAX_VP1_P1
 import no.ssb.metadata.vardef.utils.RENDERED_VARIABLE_DEFINITION_NULL_CONTACT
